requirements.txt
pydantic
fastapi[standard]
transformers
datasets 
sentencepiece
torch
langchain
textract
pdfplumber
Beautifulsoup4
pytest
pytest-asyncio
locust
boto3
pip==23
together
sqlmodel
passlib[bcrypt]
nltk
langchain-community
tqdm
pgvector
psycopg2-binary
django
djangorestframework
django-cors-headers